+++ test/orchard/inspect_test.clj
@@ -77,6 +77,12 @@
   (take-while #(not (and (string? %)
                          (re-matches #".*---.*" %))) rendered))
 
+(defn- labeled-value [label rendered]
+  (let [formatted-label (str label ": ")]
+    (->> rendered
+         (drop-while #(not (= formatted-label %)))
+         (take 2))))
+
 (defn- page-size-info [rendered]
   (let [s (last (butlast rendered))]
     (when (and (string? s) (re-find #"Page size:" s))
@@ -96,6 +102,9 @@
 
 (defn set-page-size [inspector new-size]
   (inspect/refresh inspector {:page-size new-size}))
+
+(defn set-pretty-print [inspector pretty-print]
+  (inspect/refresh inspector {:pretty-print pretty-print}))
 
 (deftest nil-test
   (testing "nil renders correctly"
@@ -1558,6 +1567,157 @@
             "  | " [:value "8" pos?] " | " [:value "8" pos?] " | " [:value "8" pos?] " | " [:newline]
             [:newline]]
            (section "Contents" rendered)))))
+
+(deftest pretty-print-map-test
+  (testing "in :pretty view-mode are pretty printed"
+    (let [rendered (-> {:a 0
+                        :bb "000"
+                        :ccc []
+                        :d [{:a 0 :bb "000" :ccc [[]]}
+                            {:a -1 :bb "111" :ccc [1]}
+                            {:a 2 :bb "222" :ccc [1 2]}]}
+                       (inspect/start)
+                       (set-pretty-print true)
+                       render)]
+      (is+ ["--- Contents:" [:newline] "  "
+            [:value ":a" 1] " = " [:value "0" 2]
+            [:newline] "  "
+            [:value ":bb" 3] " = " [:value "\"000\"" 4]
+            [:newline] "  "
+            [:value ":ccc" 5] " = " [:value "[]" 6]
+            [:newline] "  "
+            [:value ":d" 7] " = "
+            [:value (str "[{:a 0, :bb \"000\", :ccc [[]]}\n"
+                         "        {:a -1, :bb \"111\", :ccc [1]}\n"
+                         "        {:a 2, :bb \"222\", :ccc [1 2]}]") 8]
+            [:newline]]
+           (section "Contents" rendered))
+      (is (nil? (section "View mode" rendered))))))
+
+(deftest pretty-print-map-in-object-view-test
+  (testing "in :pretty view-mode are pretty printed"
+    (let [rendered (-> {:a 0
+                        :bb "000"
+                        :ccc []
+                        :d [{:a 0 :bb "000" :ccc [[]]}
+                            {:a -1 :bb "111" :ccc [1]}
+                            {:a 2 :bb "222" :ccc [1 2]}]}
+                       (inspect/start)
+                       (inspect/set-view-mode :object)
+                       (set-pretty-print true)
+                       render)]
+      (is+ ["Value: "
+            [:value (str "{:a 0,\n"
+                         "        :bb \"000\",\n"
+                         "        :ccc [],\n"
+                         "        :d\n"
+                         "        [{:a 0, :bb \"000\", :ccc [[]]}\n"
+                         "         {:a -1, :bb \"111\", :ccc [1]}\n"
+                         "         {:a 2, :bb \"222\", :ccc [1 2]}]}") 1]]
+           (labeled-value "Value" rendered))
+      (is+ ["--- View mode:" [:newline] "  :object"]
+           (section "View mode" rendered)))))
+
+(deftest pretty-print-seq-of-maps-test
+  (testing "in :pretty view-mode maps seqs of maps are pretty printed"
+    (let [rendered (-> (for [i (range 2)]
+                         {:a (- i)
+                          :bb (str i i i)
+                          :ccc (range i 0 -1)
+                          :d (for [i (range 5)]
+                               {:a (- i)
+                                :bb (str i i i)
+                                :ccc (range i 0 -1)})})
+                       (inspect/start)
+                       (set-pretty-print true)
+                       render)]
+      (is+ ["--- Contents:" [:newline]
+            "  0. "
+            [:value (str "{:a 0,\n      :bb \"000\",\n      :ccc (),\n      "
+                         ":d\n      ({:a 0, :bb \"000\", :ccc ()}\n       "
+                         "{:a -1, :bb \"111\", :ccc (1)}\n       {:a -2, :bb "
+                         "\"222\", :ccc (2 1)}\n       {:a -3, :bb \"333\", "
+                         ":ccc (3 2 1)}\n       {:a -4, :bb \"444\", :ccc "
+                         "(4 3 2 1)})}") 1]
+            [:newline]
+            "  1. "
+            [:value (str "{:a -1,\n      :bb \"111\",\n      :ccc (1),\n      "
+                         ":d\n      ({:a 0, :bb \"000\", :ccc ()}\n       "
+                         "{:a -1, :bb \"111\", :ccc (1)}\n       {:a -2, :bb "
+                         "\"222\", :ccc (2 1)}\n       {:a -3, :bb \"333\", "
+                         ":ccc (3 2 1)}\n       {:a -4, :bb \"444\", "
+                         ":ccc (4 3 2 1)})}") 2]
+            [:newline]]
+           (section "Contents" rendered))
+      (is (nil? (section "View mode" rendered))))))
+
+(deftest pretty-print-map-as-key-test
+  (testing "in :pretty view-mode maps that contain maps as a keys are pretty printed"
+    (let [rendered (-> {{:a 0
+                         :bb "000"
+                         :ccc []
+                         :d [{:a 0 :bb "000" :ccc []}
+                             {:a -1 :bb "111" :ccc [1]}
+                             {:a -2 :bb "222" :ccc [2 1]}
+                             {:a -3 :bb "333" :ccc [3 2 1]}
+                             {:a -4 :bb "444" :ccc [4 3 2 1]}]}
+                        {:a -1
+                         :bb "111"
+                         :ccc [1]
+                         :d [{:a 0 :bb "000" :ccc []}
+                             {:a -1 :bb "111" :ccc [1]}
+                             {:a -2 :bb "222" :ccc [2 1]}
+                             {:a -3 :bb "333" :ccc [3 2 1]}
+                             {:a -4 :bb "444" :ccc [4 3 2 1]}]}}
+                       (inspect/start)
+                       (set-pretty-print true)
+                       render)]
+      (is+ ["--- Contents:" [:newline] "  "
+            [:value (str "{:a 0,\n   :bb \"000\",\n   :ccc [],\n   :d\n   "
+                         "[{:a 0, :bb \"000\", :ccc []}\n    {:a -1, "
+                         ":bb \"111\", :ccc [1]}\n    {:a -2, :bb \"222\", "
+                         ":ccc [2 1]}\n    {:a -3, :bb \"333\", :ccc [3 2 1]}"
+                         "\n    {:a -4, :bb \"444\", :ccc [4 3 2 1]}]}") 1]
+            [:newline] "  =" [:newline] "  "
+            [:value (str "{:a -1,\n   :bb \"111\",\n   :ccc [1],\n   "
+                         ":d\n   [{:a 0, :bb \"000\", :ccc []}\n    "
+                         "{:a -1, :bb \"111\", :ccc [1]}\n    {:a -2, "
+                         ":bb \"222\", :ccc [2 1]}\n    {:a -3, :bb "
+                         "\"333\", :ccc [3 2 1]}\n    {:a -4, :bb "
+                         "\"444\", :ccc [4 3 2 1]}]}") 2]
+            [:newline] [:newline]]
+           (section "Contents" rendered))
+      (is (nil? (section "View mode" rendered))))))
+
+(deftest pretty-print-seq-of-map-as-key-test
+  (testing "in :pretty view-mode maps that contain seq of maps as a keys are pretty printed"
+    (let [rendered (-> {[{:a 0
+                          :bb "000"
+                          :ccc []
+                          :d [{:a 0 :bb "000" :ccc [[]]}
+                              {:a -1 :bb "111" :ccc [1]}
+                              {:a 2 :bb "222" :ccc [1 2]}]}]
+                        {:a 0
+                         :bb "000"
+                         :ccc []
+                         :d [{:a 0 :bb "000" :ccc [[]]}
+                             {:a -1 :bb "111" :ccc [1]}
+                             {:a 2 :bb "222" :ccc [1 2]}]}}
+                       (inspect/start)
+                       (set-pretty-print true)
+                       render)]
+      (is+ ["--- Contents:" [:newline] "  "
+            [:value (str "[{:a 0,\n    :bb \"000\",\n    :ccc [],\n    :d\n    "
+                         "[{:a 0, :bb \"000\", :ccc [[]]}\n     {:a -1, :bb \"111\", "
+                         ":ccc [1]}\n     {:a 2, :bb \"222\", :ccc [1 2]}]}]") 1]
+            [:newline] "  =" [:newline] "  "
+            [:value (str "{:a 0,\n   :bb \"000\",\n   :ccc [],\n   :d\n   "
+                         "[{:a 0, :bb \"000\", :ccc [[]]}\n    {:a -1, "
+                         ":bb \"111\", :ccc [1]}\n    {:a 2, :bb \"222\", "
+                         ":ccc [1 2]}]}") 2]
+            [:newline] [:newline]]
+           (section "Contents" rendered))
+      (is (nil? (section "View mode" rendered))))))
 
 (deftest tap-test
   (testing "tap-current-value"