+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master (unreleased)
 
+* [#335](https://github.com/clojure-emacs/orchard/pull/335) Add `orchard.pp` and pretty view mode.
+
 ## 0.33.0 (2025-04-08)
 
 * [#333](https://github.com/clojure-emacs/orchard/pull/333): Add `orchard.profile`.